fix: consolidate shared Wrapper component and remove duplicate debounce in EnsName demo
