+++ src/components/pageComponents/home/Examples/demos/TransactionButton/ERC20ApproveAndTransferButton/index.tsx
@@ -2,7 +2,7 @@ import { type Abi, type Address, formatUnits } from 'viem'
 import { baseSepolia } from 'viem/chains'
 import BaseERC20ApproveAndTransferButton from '@/src/components/pageComponents/home/Examples/demos/TransactionButton/ERC20ApproveAndTransferButton/ERC20ApproveAndTransferButton'
 import MintUSDC from '@/src/components/pageComponents/home/Examples/demos/TransactionButton/ERC20ApproveAndTransferButton/MintUSDC'
-import Wrapper from '@/src/components/pageComponents/home/Examples/demos/TransactionButton/Wrapper'
+import Wrapper from '@/src/components/pageComponents/home/Examples/wrapper'
 import { useSuspenseReadErc20BalanceOf } from '@/src/contracts/generated'
 import { formatNumberOrString, NumberType, withSuspense } from '@/src/core/utils'
 import type { TransactionParams } from '@/src/sdk/core'
