+++ src/components/pageComponents/home/Examples/demos/TransactionButton/NativeToken.tsx
@@ -3,7 +3,7 @@ import { type ReactElement, useState } from 'react'
 import type { Address, TransactionReceipt } from 'viem'
 import { parseEther } from 'viem'
 import { baseSepolia } from 'viem/chains'
-import Wrapper from '@/src/components/pageComponents/home/Examples/demos/TransactionButton/Wrapper'
+import Wrapper from '@/src/components/pageComponents/home/Examples/wrapper'
 import { GeneralMessage, PrimaryButton } from '@/src/core/components'
 import type { TransactionParams, TransactionResult } from '@/src/sdk/core'
 import type { EvmRawTransaction } from '@/src/sdk/core/evm/types'
